# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PredSea MVP API
  description: File-backed API for PredSea route evidence, briefings, and captain
    questions.
  version: 0.1.0
paths:
  /warnings/active:
    post:
      tags:
      - warnings
      summary: Register pre-calculated active anomaly or official warnings
      description: Allows ingestion pipelines or scripts to push pre-calculated active
        warning items directly to the API.
      operationId: post_warnings_active_endpoint_warnings_active_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/WarningItem'
              title: Warnings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Post Warnings Active Endpoint Warnings Active Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - warnings
      summary: Operational warnings from official alerts and anomaly detection
      description: Combine AEMET official CAP alerts with PredSea anomaly warnings
        computed from the latest evidence rows and climatology baseline.
      operationId: warnings_active_endpoint_warnings_active_get
      parameters:
      - name: route
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional route id, e.g. palma_ibiza
          title: Route
        description: Optional route id, e.g. palma_ibiza
      - name: place
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional place id, e.g. palma
          title: Place
        description: Optional place id, e.g. palma
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional ISO date YYYY-MM-DD
          title: Date
        description: Optional ISO date YYYY-MM-DD
      - name: z_threshold
        in: query
        required: false
        schema:
          type: number
          minimum: 0.0
          default: 1.5
          title: Z Threshold
      - name: lookback_hours
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 240
          title: Lookback Hours
      - name: min_window_hours
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 240
          title: Min Window Hours
      - name: min_sample_count
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 10
          title: Min Sample Count
      - name: include_aemet
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Aemet
      - name: include_anomaly
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Anomaly
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarningsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WarningItem:
      properties:
        source:
          type: string
          title: Source
        severity:
          type: string
          enum:
          - severe
          - moderate
          - info
          title: Severity
        variable:
          anyOf:
          - type: string
          - type: 'null'
          title: Variable
        label:
          type: string
          title: Label
        description:
          type: string
          title: Description
        value:
          anyOf:
          - type: number
          - type: 'null'
          title: Value
        unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit
        z_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Z Score
        baseline_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Baseline Type
        station_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Station Id
        station_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Station Name
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        issued_at_utc:
          type: string
          title: Issued At Utc
        valid_from_utc:
          anyOf:
          - type: string
          - type: 'null'
          title: Valid From Utc
        valid_to_utc:
          anyOf:
          - type: string
          - type: 'null'
          title: Valid To Utc
        route:
          anyOf:
          - type: string
          - type: 'null'
          title: Route
        aemet_event:
          anyOf:
          - type: string
          - type: 'null'
          title: Aemet Event
        aemet_area:
          anyOf:
          - type: string
          - type: 'null'
          title: Aemet Area
        extra:
          additionalProperties: true
          type: object
          title: Extra
      type: object
      required:
      - source
      - severity
      - label
      - description
      - issued_at_utc
      title: WarningItem
    WarningsContext:
      properties:
        route:
          anyOf:
          - type: string
          - type: 'null'
          title: Route
        place:
          anyOf:
          - type: string
          - type: 'null'
          title: Place
        date:
          anyOf:
          - type: string
          - type: 'null'
          title: Date
        lookback_hours:
          type: integer
          title: Lookback Hours
          default: 240
        min_window_hours:
          type: integer
          title: Min Window Hours
          default: 240
        min_sample_count:
          type: integer
          title: Min Sample Count
          default: 10
        z_threshold:
          type: number
          title: Z Threshold
          default: 1.5
      type: object
      title: WarningsContext
    WarningsResponse:
      properties:
        generated_at_utc:
          type: string
          title: Generated At Utc
        context:
          $ref: '#/components/schemas/WarningsContext'
        summary:
          $ref: '#/components/schemas/WarningsSummary'
        operational_stance:
          type: string
          title: Operational Stance
        warnings:
          items:
            $ref: '#/components/schemas/WarningItem'
          type: array
          title: Warnings
        sources_available:
          items:
            type: string
          type: array
          title: Sources Available
      type: object
      required:
      - generated_at_utc
      - context
      - summary
      - operational_stance
      title: WarningsResponse
    WarningsSummary:
      properties:
        total:
          type: integer
          title: Total
          default: 0
        severe:
          type: integer
          title: Severe
          default: 0
        moderate:
          type: integer
          title: Moderate
          default: 0
        info:
          type: integer
          title: Info
          default: 0
        aemet_official:
          type: integer
          title: Aemet Official
          default: 0
        predsea_anomaly:
          type: integer
          title: Predsea Anomaly
          default: 0
        sources_available:
          items:
            type: string
          type: array
          title: Sources Available
      type: object
      title: WarningsSummary